Catch NaN and do NOT try and make pay. (#2118)
Some users are seeing an old version of this script and the new version of the site. This catches the flaw where they mismatch and skips payment if it cannot determine a number to pay. We also need to make sure that https://elementary.io/scripts/download.js accurately reflects this change.

This pull request is ready for review.